Safari-611.1.10/LayoutTests/fast/sub-pixel/vertical-align-middle-overflow-expected.txt

PASS Does not overflow.
PASS Does not overflow.
PASS Does not overflow.
  text
  text
  text
The blocks above should not have scrollbars.
